IPython is now optional
