+++ requirements.txt
@@ -1,3 +1,4 @@
 -r requirements/runtime.txt
+-r requirements/ipython.txt
 -r requirements/build.txt
 -r requirements/tests.txt
