--- requirements/runtime.txt
@@ -1,2 +1,0 @@
-IPython >=0.13 ; python_version >= '3.7'
-IPython >=0.13, <7.17.0 ; python_version <= '3.6'
